+++ settings.gradle.kts
@@ -1,7 +1,7 @@
 rootProject.name = "federation-jvm"
 
 include(":federation-graphql-java-support")
-include(":federation-subscription-callback")
+include(":federation-spring-subscription-callback")
 
 project(":federation-graphql-java-support").projectDir = file("graphql-java-support")
-project(":federation-subscription-callback").projectDir = file("subscription-callback")
+project(":federation-spring-subscription-callback").projectDir = file("spring-subscription-callback")
