+++ spring-subscription-callback/build.gradle.kts
