+++ spring-subscription-callback/src/main/java/com/apollographql/subscription/exception/InactiveSubscriptionException.java
@@ -2,6 +2,10 @@ package com.apollographql.subscription.exception;
 
 import com.apollographql.subscription.callback.SubscriptionCallback;
 
+/**
+ * Exception thrown when subscription becomes inactive (router responds using 2xx status to any
+ * callback message).
+ */
 public class InactiveSubscriptionException extends RuntimeException {
 
   public InactiveSubscriptionException(SubscriptionCallback callback) {
