+++ spring-subscription-callback/src/main/java/com/apollographql/subscription/exception/InvalidCallbackExtensionException.java
