+++ spring-subscription-callback/src/main/java/com/apollographql/subscription/message/CallbackMessageAction.java
@@ -2,11 +2,10 @@ package com.apollographql.subscription.message;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+/** HTTP callback message type */
 public enum CallbackMessageAction {
   @JsonProperty("check")
   CHECK,
-  @JsonProperty("heartbeat")
-  HEARTBEAT,
   @JsonProperty("next")
   NEXT,
   @JsonProperty("complete")
