+++ spring-subscription-callback/src/main/java/com/apollographql/subscription/message/CallbackMessageComplete.java
@@ -6,16 +6,22 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+/**
+ * <code>complete</code> message is send to the <strong>Router</strong> to terminate active
+ * subscription. Subscription can terminate either by reaching end of the data stream OR by
+ * encountering an error that caused subscription to fail.<br>
+ * If subscription failed due to an error, <code>complete</code> message should include list of
+ * <code>GraphQLError</code>s that caused the failure.
+ *
+ * @param id unique subscription ID
+ * @param verifier value provided by Router that is used to validate requests
+ * @param errors optional list of errors if subscription terminated abnormally
+ */
 public record CallbackMessageComplete(String id, String verifier, List<GraphQLError> errors)
     implements SubscritionCallbackMessage {
 
   public CallbackMessageComplete(String id, String verifier) {
     this(id, verifier, Collections.emptyList());
-  }
-
-  @Override
-  public String getKind() {
-    return "subscription";
   }
 
   @Override
