+++ spring-subscription-callback/src/main/java/com/apollographql/subscription/message/SubscritionCallbackMessage.java
@@ -1,8 +1,11 @@
 package com.apollographql.subscription.message;
 
+/** Common interface for HTTP callback messages. */
 public sealed interface SubscritionCallbackMessage
     permits CallbackMessageCheck, CallbackMessageNext, CallbackMessageComplete {
-  String getKind(); //  default "subscription";
+  default String getKind() {
+    return "subscription";
+  }
 
   CallbackMessageAction getAction();
