+++ spring-subscription-callback/src/main/java/com/apollographql/subscription/webflux/CallbackGraphQlHttpHandler.java
@@ -21,7 +21,15 @@ import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
 import reactor.core.publisher.Mono;
 
-/** Copy of org.springframework.graphql.server.webflux.GraphQlHttpHandler */
+/**
+ * Functional GraphQL handler for reactive WebFlux applications that supports Apollo Subscription
+ * Callback Protocol.
+ *
+ * @see <a
+ *     href="https://www.apollographql.com/docs/router/executing-operations/subscription-callback-protocol">Subscription
+ *     Callback Protocol</a>
+ * @see org.springframework.graphql.server.webflux.GraphQlHttpHandler
+ */
 public class CallbackGraphQlHttpHandler extends GraphQlHttpHandler {
 
   private static final Log logger = LogFactory.getLog(GraphQlHttpHandler.class);
@@ -80,7 +88,7 @@ public class CallbackGraphQlHttpHandler extends GraphQlHttpHandler {
                     .flatMap(
                         callback ->
                             this.subscriptionCallbackHandler
-                                .handleCallback(graphQlRequest, callback)
+                                .handleSubscriptionUsingCallback(graphQlRequest, callback)
                                 .flatMap(
                                     success -> {
                                       if (success) {
