+++ spring-subscription-callback/src/test/java/com/apollographql/subscription/configuration/TestGraphQLConfiguration.java
