+++ spring-subscription-callback/src/test/resources/graphql/schema.graphqls
