--- subscription-callback/README.md
@@ -1,1 +1,0 @@
-# Subscription Callback Support for Spring GraphQL
