--- subscription-callback/src/main/java/com/apollographql/subscription/message/CallbackMessageCheck.java
@@ -1,25 +1,0 @@
-package com.apollographql.subscription.message;
-
-public record CallbackMessageCheck(String id, String verifier)
-    implements SubscritionCallbackMessage {
-
-  @Override
-  public String getKind() {
-    return "subscription";
-  }
-
-  @Override
-  public CallbackMessageAction getAction() {
-    return CallbackMessageAction.CHECK;
-  }
-
-  @Override
-  public String getId() {
-    return id;
-  }
-
-  @Override
-  public String getVerifier() {
-    return verifier;
-  }
-}
